requirements.txt
python==3.11
SQLAlchemy==2.0.0			# SQLAlchemy for ORM
psycopg2==2.9.3		        # PostgreSQL adapter for SQLAlchemy
Click==8.1.3				# Click for creating command-line interfaces
pytest==7.2.2				# Pytest for testing
Sphinx==5.3.0				# Sphinx for generating documentation
pyzmq==25.1.1				# PyZMQ for ZeroMQ bindings in Python
python-dotenv